+++ CHANGELOG.md
@@ -7,6 +7,7 @@ See the [keepachangelog.com description](https://keepachangelog.com/en/1.0.0/).
 * Added
   * String::replaceChar
   * TextBox - param scrollLines
+  * WindowStateBase::forEachChild
 
 ## 0.23.0 - 2025-04-14